rename: SignUp 타입 코로케이션 적용
