+++ src/pages/SignUp/types/index.ts
@@ -1,5 +1,3 @@
-import { DetailSkillOption } from '../../../types/signUp';
-
 export interface DropdownValue {
   mainSkill: string;
   skillDepthOne: string;
@@ -10,4 +8,47 @@ export interface DropdownValue {
 
 export interface DetailSkills {
   [key: number]: DetailSkillOption[];
+}
+
+export interface PostSignUp {
+  nickname: string;
+  mainSkillId: number;
+  profileImageUrl: string;
+  skills: {
+    skillId: number;
+    level: string;
+  }[];
+  joinPurposes: number[];
+  livingPlace?: string;
+  workingPlace?: string;
+  introduction?: string;
+  email: string;
+  oauthId: string;
+  oauthServerType: string;
+}
+
+export interface Skill {
+  id: number;
+  name: string;
+}
+
+export interface GetSignUp {
+  mainSkillResponses: MainSkillOption[];
+  purposeResponses: CheckboxOption[];
+}
+
+export interface MainSkillOption {
+  id: number;
+  name: string;
+  detailSkillResponses: DetailSkillOption[];
+}
+
+export interface DetailSkillOption {
+  id: number;
+  name: string;
+}
+
+export interface CheckboxOption {
+  id: number;
+  name: string;
 }
